Create control message broker file + cr
